DBZ-1478 Remove Thorntail based order-service
The docker-compose.yaml defaults for QUARKUS_BUILD were changed to favor JVM mode
as the default to adhere to the same defaults as the maven build setup.  A user
must explicitly set QUARKUS_BUILD=native to build native image containers.

The README.md was updated to be a more informative and to align with all changes
made in this commit.