[KStreams] Fix build errors
When using newer versions of Maven where HTTPS repository access
is required and non-secure access is blocked, the project failed
to build the artifacts from the sibling repository. In addition,
moving to JDK11 in the Maven POM required that the build step in
docker install JDK11-based Maven tooling rather than JDK8.